refactor api controllers to match class names to api paths
