+++ src/main/java/org/trebol/api/RootController.java
@@ -1,4 +1,4 @@
-package org.trebol.api.controllers;
+package org.trebol.api;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
  * @author Benjamin La Madrid <bg.lamadrid@gmail.com>
  */
 @RestController
-public class DefaultController {
+public class RootController {
 
   @GetMapping({"", "/"})
   public ResponseEntity defaultMapping() {
