+++ src/main/java/org/trebol/api/controllers/DataUserRolesController.java
@@ -1,4 +1,4 @@
-package org.trebol.api.controllers.data;
+package org.trebol.api.controllers;
 
 import java.util.Collection;
 import java.util.Map;
@@ -22,57 +22,57 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import org.trebol.api.GenericCrudController;
-import org.trebol.api.pojo.ProductTypePojo;
+import org.trebol.api.pojo.UserRolePojo;
 import org.trebol.config.CustomProperties;
-import org.trebol.jpa.entities.ProductType;
+import org.trebol.jpa.entities.UserRole;
 import org.trebol.jpa.services.GenericCrudService;
 
 /**
- * API point of entry for ProductType entities
+ * API point of entry for UserRole entities
  *
  * @author Benjamin La Madrid <bg.lamadrid at gmail.com>
  */
 @RestController
-@RequestMapping("/data/product_types")
-public class ProductTypesDataController
-    extends GenericCrudController<ProductTypePojo, ProductType, Integer> {
+@RequestMapping("/data/user_roles")
+public class DataUserRolesController
+    extends GenericCrudController<UserRolePojo, UserRole, Integer> {
 
   @Autowired
-  public ProductTypesDataController(CustomProperties globals,
-      GenericCrudService<ProductTypePojo, ProductType, Integer> crudService) {
+  public DataUserRolesController(CustomProperties globals,
+      GenericCrudService<UserRolePojo, UserRole, Integer> crudService) {
     super(globals, crudService);
   }
 
-  @GetMapping
-  @PreAuthorize("hasAuthority('product_types:read')")
-  public Collection<ProductTypePojo> readMany(@RequestParam Map<String, String> allRequestParams) {
+  @GetMapping({"", "/"})
+  @PreAuthorize("hasAuthority('user_roles:read')")
+  public Collection<UserRolePojo> readMany(@RequestParam Map<String, String> allRequestParams) {
     return super.readMany(null, null, allRequestParams);
   }
 
   @Override
-  @PostMapping
-  @PreAuthorize("hasAuthority('product_types:create')")
-  public Integer create(@RequestBody @Valid ProductTypePojo input) {
+  @PostMapping({"", "/"})
+  @PreAuthorize("hasAuthority('user_roles:create')")
+  public Integer create(@RequestBody @Valid UserRolePojo input) {
     return super.create(input);
   }
 
   @Override
-  @GetMapping("/{id}")
-  @PreAuthorize("hasAuthority('product_types:read')")
-  public ProductTypePojo readOne(@PathVariable Integer id) {
+  @GetMapping({"/{id}", "/{id}/"})
+  @PreAuthorize("hasAuthority('user_roles:read')")
+  public UserRolePojo readOne(@PathVariable Integer id) {
     return super.readOne(id);
   }
 
   @Override
-  @PutMapping("/{id}")
-  @PreAuthorize("hasAuthority('product_types:update')")
-  public Integer update(@RequestBody @Valid ProductTypePojo input, @PathVariable Integer id) {
+  @PutMapping({"/{id}", "/{id}/"})
+  @PreAuthorize("hasAuthority('user_roles:update')")
+  public Integer update(@RequestBody @Valid UserRolePojo input, @PathVariable Integer id) {
     return super.update(input, id);
   }
 
   @Override
-  @DeleteMapping("/{id}")
-  @PreAuthorize("hasAuthority('product_types:delete')")
+  @DeleteMapping({"/{id}", "/{id}/"})
+  @PreAuthorize("hasAuthority('user_roles:delete')")
   public boolean delete(@PathVariable Integer id) {
     return super.delete(id);
   }
