+++ src/main/java/org/trebol/api/controllers/DataUsersController.java
@@ -1,4 +1,4 @@
-package org.trebol.api.controllers.data.management;
+package org.trebol.api.controllers;
 
 import java.util.Collection;
 import java.util.Map;
@@ -22,57 +22,57 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import org.trebol.api.GenericCrudController;
-import org.trebol.api.pojo.CustomerPojo;
+import org.trebol.api.pojo.UserPojo;
 import org.trebol.config.CustomProperties;
-import org.trebol.jpa.entities.Customer;
+import org.trebol.jpa.entities.User;
 import org.trebol.jpa.services.GenericCrudService;
 
 /**
- * API point of entry for Customer entities
+ * API point of entry for User entities
  *
  * @author Benjamin La Madrid <bg.lamadrid at gmail.com>
  */
 @RestController
-@RequestMapping("/data/customers")
-public class CustomersDataManagementController
-    extends GenericCrudController<CustomerPojo, Customer, Integer> {
+@RequestMapping("/data/users")
+public class DataUsersController
+    extends GenericCrudController<UserPojo, User, Integer> {
 
   @Autowired
-  public CustomersDataManagementController(CustomProperties globals,
-      GenericCrudService<CustomerPojo, Customer, Integer> crudService) {
+  public DataUsersController(CustomProperties globals,
+      GenericCrudService<UserPojo, User, Integer> crudService) {
     super(globals, crudService);
   }
 
-  @GetMapping
-  @PreAuthorize("hasAuthority('customers:read')")
-  public Collection<CustomerPojo> readMany(@RequestParam Map<String, String> allRequestParams) {
+  @GetMapping({"", "/"})
+  @PreAuthorize("hasAuthority('users:read')")
+  public Collection<UserPojo> readMany(@RequestParam Map<String, String> allRequestParams) {
     return super.readMany(null, null, allRequestParams);
   }
 
   @Override
-  @PostMapping
-  @PreAuthorize("hasAuthority('customers:create')")
-  public Integer create(@RequestBody @Valid CustomerPojo input) {
+  @PostMapping({"", "/"})
+  @PreAuthorize("hasAuthority('users:create')")
+  public Integer create(@RequestBody @Valid UserPojo input) {
     return super.create(input);
   }
 
   @Override
-  @GetMapping("/{id}")
-  @PreAuthorize("hasAuthority('customers:read')")
-  public CustomerPojo readOne(@PathVariable Integer id) {
+  @GetMapping({"/{id}", "/{id}/"})
+  @PreAuthorize("hasAuthority('users:read')")
+  public UserPojo readOne(@PathVariable Integer id) {
     return super.readOne(id);
   }
 
   @Override
-  @PutMapping("/{id}")
-  @PreAuthorize("hasAuthority('customers:update')")
-  public Integer update(@RequestBody @Valid CustomerPojo input, @PathVariable Integer id) {
+  @PutMapping({"/{id}", "/{id}/"})
+  @PreAuthorize("hasAuthority('users:update')")
+  public Integer update(@RequestBody @Valid UserPojo input, @PathVariable Integer id) {
     return super.update(input, id);
   }
 
   @Override
-  @DeleteMapping("/{id}")
-  @PreAuthorize("hasAuthority('customers:delete')")
+  @DeleteMapping({"/{id}", "/{id}/"})
+  @PreAuthorize("hasAuthority('users:delete')")
   public boolean delete(@PathVariable Integer id) {
     return super.delete(id);
   }
